aula05-orchestration-airflow/requirements.txt
apache-airflow>=2.8
